AERO3760_A1_Scaffold_Code/aero3760_requirements.txt
numpy==2.3.2
scipy==1.16.0
pandas==2.3.1
matplotlib==3.10.3
seaborn==0.13.2
pyvista==0.43.6
PyQt6==6.9.1

spacetools-0.1.0-py3-none-any.whl